Done api server unit test
